Schritt 0: PSRAM-Anzeige korrigiert, Touch im Log, Hinweis zum Monitor-Filter
- PSRAM mit Nachkommastelle statt abgerundet (7.9 statt 7 MB)
- TOUCH down/up mit Koordinaten im seriellen Log für Ferndiagnose
- TESTEN.md: Hinweis, dass die Debug-Meldung des Exception-Decoders kein Fehler ist

+++ src/main.cpp
@@ -28,7 +28,8 @@ void logBootInfo() {
     Serial.printf("Chip: %s Rev %d, %d Kerne, %lu MHz\n", ESP.getChipModel(), ESP.getChipRevision(),
                   ESP.getChipCores(), static_cast<unsigned long>(ESP.getCpuFreqMHz()));
     Serial.printf("Flash: %lu MB\n", static_cast<unsigned long>(ESP.getFlashChipSize() / (1024 * 1024)));
-    Serial.printf("PSRAM: %lu MB\n", static_cast<unsigned long>(ESP.getPsramSize() / (1024 * 1024)));
+    // Ein kleiner Teil des 8-MB-PSRAM ist reserviert, daher mit Nachkommastelle statt abgerundet.
+    Serial.printf("PSRAM: %.1f MB\n", ESP.getPsramSize() / (1024.0 * 1024.0));
 }
 
 void logStatus(uint32_t now) {
